+++ entity/src/tx_credential.rs
@@ -4,10 +4,12 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Deserialize, Serialize)]
 #[sea_orm(table_name = "TxCredentialRelation")]
 pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i32,
-    pub credential_id: i32,
-    pub tx_id: i32,
+    #[sea_orm(primary_key, column_type = "BigInteger")]
+    pub id: i64,
+    #[sea_orm(column_type = "BigInteger")]
+    pub credential_id: i64,
+    #[sea_orm(column_type = "BigInteger")]
+    pub tx_id: i64,
     pub relation: i32,
 }
